+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 20.07.2018
+
+* [Rust] Code of Kutulu - Handle explorer with if let (CodinGame)
+
 #### 19.07.2018
 
 * [Rust] Code of Kutulu - Fixed distance unwrap (CodinGame)